[11.0][MIG] ursa_product_reference
